CL-Project/python-microservice/core/craft/requirements.txt
paddlepaddle-gpu==2.6.1
# paddlepaddle-gpu==2.5.2
paddleocr==2.6.1.3
paddlex==3.1.1
# --extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.2 # +cu118
torchvision==0.17.2
torchaudio==2.2.2
transformers==4.46.0
sentencepiece==0.1.99
easyocr==1.7.1
pix2text==1.1.3.2
cnocr==2.3.2.1
cnstd==1.2.6
rapidocr-onnxruntime==1.4.4
onnxruntime-gpu==1.17.1
opencv-python-headless==4.10.0.82
Pillow==10.3.0
scikit-image==0.23.2
numpy==1.26.4
tqdm==4.66.4
rapidfuzz==3.6.2
ftfy==6.1.3
shapely==1.8.5
nvidia-cudnn-cu11==8.9.5.29
